refactor: split of GuiManager code into seperate Panel-objects
