+++ webgpu_app/ImGuiManager.cpp
@@ -1,0 +1,229 @@
+/*****************************************************************************
+ * weBIGeo
+ * Copyright (C) 2026 Gerald Kimmersdorfer
+ * Copyright (C) 2025 Patrick Komon
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *****************************************************************************/
+
+#include "ImGuiManager.h"
+#include "TerrainRenderer.h"
+#include "RenderingContext.h"
+
+#include "imgui/ImGuiPanel.h"
+
+#ifdef ALP_WEBGPU_APP_ENABLE_IMGUI
+#include "imgui/TimingPanel.h"
+#include "imgui/CameraPanel.h"
+#include "imgui/AppPanel.h"
+#include "imgui/CloudPanel.h"
+#include "imgui/EnginePanel.h"
+#include "imgui/AboutPanel.h"
+#include "imgui/CompassPanel.h"
+#include "imgui/LogoPanel.h"
+#include <imgui_internal.h>
+#include "backends/imgui_impl_sdl2.h"
+#include "backends/imgui_impl_wgpu.h"
+#include <IconsFontAwesome5.h>
+#include <imnodes.h>
+#endif
+
+#include "util/dark_mode.h"
+#include <QDebug>
+#include <QFile>
+
+namespace webgpu_app {
+
+ImGuiManager::ImGuiManager(TerrainRenderer* terrain_renderer)
+    : m_terrain_renderer(terrain_renderer)
+{
+}
+
+void ImGuiManager::init(
+    SDL_Window* window, WGPUDevice device, [[maybe_unused]] WGPUTextureFormat swapchainFormat, [[maybe_unused]] WGPUTextureFormat depthTextureFormat)
+{
+    qDebug() << "Setup ImGuiManager...";
+    m_window = window;
+    m_device = device;
+
+#ifdef ALP_WEBGPU_APP_ENABLE_IMGUI
+    // Setup Dear ImGui context
+    IMGUI_CHECKVERSION();
+    ImGui::CreateContext();
+
+    // Setup ImNodes
+    ImNodes::CreateContext();
+
+    // Setup Platform/Renderer backends
+    ImGui_ImplSDL2_InitForOther(m_window);
+    ImGui_ImplWGPU_InitInfo init_info = {};
+    init_info.Device = m_device;
+    init_info.RenderTargetFormat = swapchainFormat;
+    init_info.DepthStencilFormat = depthTextureFormat;
+    init_info.NumFramesInFlight = 3;
+    ImGui_ImplWGPU_Init(&init_info);
+
+    webgpu_app::util::setup_darkmode_imgui_style();
+    install_fonts();
+
+    auto* rc = m_terrain_renderer->get_rendering_context();
+    auto* engine_ctx = rc->engine_context();
+
+    m_panels.push_back(std::make_unique<LogoPanel>(m_device));
+    m_panels.push_back(std::make_unique<CompassPanel>(m_terrain_renderer));
+    m_panels.push_back(std::make_unique<AboutPanel>());
+    m_panels.push_back(std::make_unique<TimingPanel>(m_terrain_renderer));
+    m_panels.push_back(std::make_unique<CameraPanel>(m_terrain_renderer));
+    m_panels.push_back(std::make_unique<AppPanel>(m_terrain_renderer));
+    m_panels.push_back(std::make_unique<CloudPanel>(rc->clouds_manager(), engine_ctx->cloud_geometry()));
+    m_panels.push_back(std::make_unique<EnginePanel>(m_terrain_renderer));
+#endif
+}
+
+void ImGuiManager::install_fonts()
+{
+#ifdef ALP_WEBGPU_APP_ENABLE_IMGUI
+    ImGuiIO& io = ImGui::GetIO();
+
+    float baseFontSize = 16.0f;
+    float iconFontSize = 14.0f;
+
+    {
+        QFile file(":/fonts/Roboto-Regular.ttf");
+        if (!file.open(QIODevice::ReadOnly)) {
+            throw std::runtime_error("Failed to open Main Font.");
+        }
+        QByteArray byteArray = file.readAll();
+        file.close();
+
+        ImFontConfig font_cfg;
+        font_cfg.FontDataOwnedByAtlas = false;
+        io.Fonts->AddFontFromMemoryTTF(byteArray.data(), byteArray.size(), baseFontSize, &font_cfg);
+    }
+
+    {
+        QFile file(":/fonts/fa5-solid-900.ttf");
+        if (!file.open(QIODevice::ReadOnly)) {
+            throw std::runtime_error("Failed to open glyph font.");
+        }
+        QByteArray byteArray = file.readAll();
+        file.close();
+
+        // merge in icons from Font Awesome
+        static const ImWchar icons_ranges[] = { ICON_MIN_FA, ICON_MAX_16_FA, 0 };
+        ImFontConfig icons_config;
+        icons_config.MergeMode = true;
+        icons_config.PixelSnapH = true;
+        icons_config.GlyphMinAdvanceX = iconFontSize;
+        icons_config.FontDataOwnedByAtlas = false;
+        io.Fonts->AddFontFromMemoryTTF(byteArray.data(), byteArray.size(), iconFontSize, &icons_config, icons_ranges);
+    }
+#endif
+}
+
+void ImGuiManager::render([[maybe_unused]] WGPURenderPassEncoder renderPass)
+{
+#ifdef ALP_WEBGPU_APP_ENABLE_IMGUI
+    ImGui_ImplWGPU_NewFrame();
+    ImGui_ImplSDL2_NewFrame();
+    ImGui::NewFrame();
+
+    draw();
+
+    ImGui::Render();
+    ImGui_ImplWGPU_RenderDrawData(ImGui::GetDrawData(), renderPass);
+#endif
+}
+
+void ImGuiManager::shutdown()
+{
+#ifdef ALP_WEBGPU_APP_ENABLE_IMGUI
+    qDebug() << "Releasing ImGuiManager...";
+    ImGui_ImplWGPU_Shutdown();
+    ImGui_ImplSDL2_Shutdown();
+    ImNodes::DestroyContext();
+    ImGui::DestroyContext();
+#endif
+}
+
+bool ImGuiManager::want_capture_keyboard()
+{
+#ifdef ALP_WEBGPU_APP_ENABLE_IMGUI
+    return ImGui::GetIO().WantCaptureKeyboard;
+#else
+    return false;
+#endif
+}
+
+bool ImGuiManager::want_capture_mouse()
+{
+#ifdef ALP_WEBGPU_APP_ENABLE_IMGUI
+    return ImGui::GetIO().WantCaptureMouse;
+#else
+    return false;
+#endif
+}
+
+void ImGuiManager::on_sdl_event(SDL_Event& event)
+{
+#ifdef ALP_WEBGPU_APP_ENABLE_IMGUI
+    ImGui_ImplSDL2_ProcessEvent(&event);
+#endif
+}
+
+void ImGuiManager::set_gui_visibility(bool visible) { m_gui_visible = visible; }
+
+bool ImGuiManager::get_gui_visibility() const { return m_gui_visible; }
+
+void ImGuiManager::draw()
+{
+    if (m_first_frame) {
+        // Initialize app panel zoom level
+        for (auto& panel : m_panels) {
+            if (auto* app = dynamic_cast<AppPanel*>(panel.get())) {
+                app->init();
+            }
+        }
+#ifndef QT_DEBUG
+        // Open disclaimer on first non-debug frame
+        for (auto& panel : m_panels) {
+            if (auto* about = dynamic_cast<AboutPanel*>(panel.get())) {
+                about->open_disclaimer();
+            }
+        }
+#endif
+        m_first_frame = false;
+    }
+
+#ifdef ALP_WEBGPU_APP_ENABLE_IMGUI
+    if (!m_gui_visible)
+        return;
+
+    // Standalone windows, overlays, and modals
+    for (auto& panel : m_panels)
+        panel->draw();
+
+    // Main sidebar window with CollapsingHeader sections
+    ImGui::SetNextWindowPos(ImVec2(ImGui::GetIO().DisplaySize.x - 430, 0)); // Set position to top-right corner
+    ImGui::SetNextWindowSize(ImVec2(430, ImGui::GetIO().DisplaySize.y)); // Set height to full screen height, width as desired
+    ImGui::Begin("weBIGeo", nullptr, ImGuiWindowFlags_NoMove | ImGuiWindowFlags_NoResize | ImGuiWindowFlags_NoTitleBar);
+
+    for (auto& panel : m_panels)
+        panel->draw_panel();
+
+    ImGui::End();
+#endif
+}
+
+} // namespace webgpu_app
